test(tools): split Scout tests into per-concern modules
tools/tests.rs (1024) split into five sibling test files (query_tests,
repo_lazy_tests, repo_io_tests, stdin_tests, builder_tests) plus a
test_helpers module holding the shared scout_with_brave/scout_with_github/
scout_lazy builders as pub(super) fns. All files <=400 lines; T-SB/T-DNS/T-R
IDs and fn names preserved; production tools.rs unchanged apart from the mod
declarations; 449 tests pass.
